migrate  theme storage
